Reorganize from plistgen to shortcutgen
